Update
This is what I have so far.  I think the SearchExample can work for both
searching (which right now sucks) and as an outline for how to
highlight.  The text files are some text books that I made so I'd have
something to test with, the pictures will be the icons for each book.
Have fun